# backend/api/gen/go/admin/service/v1/i_budget_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_budget.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-ledger/api/gen/go/ledger/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBudgetServiceCreate = "/admin.service.v1.BudgetService/Create"
const OperationBudgetServiceDelete = "/admin.service.v1.BudgetService/Delete"
const OperationBudgetServiceGet = "/admin.service.v1.BudgetService/Get"
const OperationBudgetServiceGetProgress = "/admin.service.v1.BudgetService/GetProgress"
const OperationBudgetServiceList = "/admin.service.v1.BudgetService/List"
const OperationBudgetServiceListAll = "/admin.service.v1.BudgetService/ListAll"
const OperationBudgetServiceUpdate = "/admin.service.v1.BudgetService/Update"

type BudgetServiceHTTPServer interface {
	Create(context.Context, *v11.CreateBudgetRequest) (*v11.Budget, error)
	Delete(context.Context, *v11.DeleteBudgetRequest) (*emptypb.Empty, error)
	Get(context.Context, *v11.GetBudgetRequest) (*v11.Budget, error)
	GetProgress(context.Context, *v11.GetBudgetProgressRequest) (*v11.BudgetProgress, error)
	List(context.Context, *v1.PagingRequest) (*v11.ListBudgetResponse, error)
	ListAll(context.Context, *v11.ListAllBudgetRequest) (*v11.ListBudgetResponse, error)
	Update(context.Context, *v11.UpdateBudgetRequest) (*v11.Budget, error)
}

func RegisterBudgetServiceHTTPServer(s *http.Server, srv BudgetServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/budgets", _BudgetService_List5_HTTP_Handler(srv))
	r.GET("/admin/v1/budgets-all", _BudgetService_ListAll3_HTTP_Handler(srv))
	r.GET("/admin/v1/budgets/{id}", _BudgetService_Get6_HTTP_Handler(srv))
	r.POST("/admin/v1/budgets", _BudgetService_Create4_HTTP_Handler(srv))
	r.PUT("/admin/v1/budgets/{id}", _BudgetService_Update4_HTTP_Handler(srv))
	r.DELETE("/admin/v1/budgets/{id}", _BudgetService_Delete4_HTTP_Handler(srv))
	r.GET("/admin/v1/budgets/{id}/progress", _BudgetService_GetProgress0_HTTP_Handler(srv))
}

func _BudgetService_List5_HTTP_Handler(srv BudgetServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBudgetServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListBudgetResponse)
		return ctx.Result(200, reply)
	}
}

func _BudgetService_ListAll3_HTTP_Handler(srv BudgetServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.ListAllBudgetRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBudgetServiceListAll)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListAll(ctx, req.(*v11.ListAllBudgetRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListBudgetResponse)
		return ctx.Result(200, reply)
	}
}

func _BudgetService_Get6_HTTP_Handler(srv BudgetServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetBudgetRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBudgetServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetBudgetRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Budget)
		return ctx.Result(200, reply)
	}
}

func _BudgetService_Create4_HTTP_Handler(srv BudgetServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateBudgetRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBudgetServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateBudgetRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Budget)
		return ctx.Result(200, reply)
	}
}

func _BudgetService_Update4_HTTP_Handler(srv BudgetServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateBudgetRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBudgetServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateBudgetRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Budget)
		return ctx.Result(200, reply)
	}
}

func _BudgetService_Delete4_HTTP_Handler(srv BudgetServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteBudgetRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBudgetServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteBudgetRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _BudgetService_GetProgress0_HTTP_Handler(srv BudgetServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetBudgetProgressRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBudgetServiceGetProgress)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetProgress(ctx, req.(*v11.GetBudgetProgressRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.BudgetProgress)
		return ctx.Result(200, reply)
	}
}

type BudgetServiceHTTPClient interface {
	Create(ctx context.Context, req *v11.CreateBudgetRequest, opts ...http.CallOption) (rsp *v11.Budget, err error)
	Delete(ctx context.Context, req *v11.DeleteBudgetRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Get(ctx context.Context, req *v11.GetBudgetRequest, opts ...http.CallOption) (rsp *v11.Budget, err error)
	GetProgress(ctx context.Context, req *v11.GetBudgetProgressRequest, opts ...http.CallOption) (rsp *v11.BudgetProgress, err error)
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListBudgetResponse, err error)
	ListAll(ctx context.Context, req *v11.ListAllBudgetRequest, opts ...http.CallOption) (rsp *v11.ListBudgetResponse, err error)
	Update(ctx context.Context, req *v11.UpdateBudgetRequest, opts ...http.CallOption) (rsp *v11.Budget, err error)
}

type BudgetServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewBudgetServiceHTTPClient(client *http.Client) BudgetServiceHTTPClient {
	return &BudgetServiceHTTPClientImpl{client}
}

func (c *BudgetServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateBudgetRequest, opts ...http.CallOption) (*v11.Budget, error) {
	var out v11.Budget
	pattern := "/admin/v1/budgets"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBudgetServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BudgetServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteBudgetRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/budgets/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBudgetServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BudgetServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetBudgetRequest, opts ...http.CallOption) (*v11.Budget, error) {
	var out v11.Budget
	pattern := "/admin/v1/budgets/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBudgetServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BudgetServiceHTTPClientImpl) GetProgress(ctx context.Context, in *v11.GetBudgetProgressRequest, opts ...http.CallOption) (*v11.BudgetProgress, error) {
	var out v11.BudgetProgress
	pattern := "/admin/v1/budgets/{id}/progress"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBudgetServiceGetProgress))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BudgetServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListBudgetResponse, error) {
	var out v11.ListBudgetResponse
	pattern := "/admin/v1/budgets"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBudgetServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BudgetServiceHTTPClientImpl) ListAll(ctx context.Context, in *v11.ListAllBudgetRequest, opts ...http.CallOption) (*v11.ListBudgetResponse, error) {
	var out v11.ListBudgetResponse
	pattern := "/admin/v1/budgets-all"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBudgetServiceListAll))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BudgetServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateBudgetRequest, opts ...http.CallOption) (*v11.Budget, error) {
	var out v11.Budget
	pattern := "/admin/v1/budgets/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBudgetServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}