# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-ledger
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-ledger/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/budgets:
        get:
            tags:
                - BudgetService
            operationId: BudgetService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.ListBudgetResponse'
        post:
            tags:
                - BudgetService
            operationId: BudgetService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ledger.service.v1.CreateBudgetRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.Budget'
...
    /app/v1/budgets/{id}:
        get:
            tags:
                - BudgetService
            operationId: BudgetService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.Budget'
        put:
            tags:
                - BudgetService
            operationId: BudgetService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ledger.service.v1.UpdateBudgetRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.Budget'
        delete:
            tags:
                - BudgetService
            operationId: BudgetService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        ledger.service.v1.Budget:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                tenantId:
                    type: integer
                    format: uint32
                bookId:
                    type: integer
                    format: uint32
                name:
                    type: string
                period:
                    enum:
                        - BUDGET_PERIOD_UNSPECIFIED
                        - BUDGET_PERIOD_MONTHLY
                        - BUDGET_PERIOD_YEARLY
                        - BUDGET_PERIOD_QUARTERLY
                        - BUDGET_PERIOD_WEEKLY
                    type: string
                    format: enum
                amount:
                    type: string
                usedAmount:
                    type: string
                categoryId:
                    type: integer
                    format: uint32
                accountId:
                    type: integer
                    format: uint32
                startDate:
                    type: string
                endDate:
                    type: string
                enable:
                    type: boolean
                notify:
                    type: boolean
                notes:
                    type: string
                createdBy:
                    type: integer
                    format: uint32
                updatedBy:
                    type: integer
                    format: uint32
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
            description: 预算
...
        ledger.service.v1.CreateBudgetRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/ledger.service.v1.Budget'
            description: 请求 - 创建预算
...
        ledger.service.v1.ListBudgetResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/ledger.service.v1.Budget'
                total:
                    type: string
            description: 回应 - 预算列表
...
        ledger.service.v1.UpdateBudgetRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/ledger.service.v1.Budget'
                updateMask:
                    type: string
                    format: field-mask
                allowMissing:
                    type: boolean
            description: 请求 - 更新预算
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...